[#14] rename: 임시로 controller 패키지에 경로 이동
